feat: implement centralized file download and cleanup utilities for packages and icons
