+++ internal/resources/packages/crud.go
@@ -5,10 +5,7 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
@@ -83,22 +80,7 @@ func create(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.
 		return diag.FromErr(fmt.Errorf("failed to create and upload Jamf Pro Package '%s': %v", resource.PackageName, err))
 	}
 
-	// Only clean up if we downloaded from web source and verify the path is what we expect
-	if regexp.MustCompile(`^(http|https)://`).MatchString(d.Get("package_file_source").(string)) {
-		if !strings.HasPrefix(localFilePath, os.TempDir()) {
-			log.Printf("[WARN] Refusing to remove file '%s' as it's not in the temporary directory: timestamp=%s",
-				localFilePath, time.Now().UTC().Format(time.RFC3339))
-			return diags
-		}
-
-		if err := os.Remove(localFilePath); err != nil {
-			log.Printf("[WARN] Failed to remove downloaded package file '%s': %v: timestamp=%s",
-				localFilePath, err, time.Now().UTC().Format(time.RFC3339))
-		} else {
-			log.Printf("[INFO] Successfully removed downloaded package file '%s': timestamp=%s",
-				localFilePath, time.Now().UTC().Format(time.RFC3339))
-		}
-	}
+	common.CleanupDownloadedPackage(d.Get("package_file_source").(string), localFilePath)
 
 	return append(diags, readNoCleanup(ctx, d, meta)...)
 }
@@ -201,22 +183,7 @@ func update(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.
 		return diag.FromErr(fmt.Errorf("failed to update Jamf Pro Package '%s' (ID: %s): %v", resource.PackageName, resourceID, err))
 	}
 
-	// Only clean up if we downloaded from web source and verify the path is what we expect
-	if regexp.MustCompile(`^(http|https)://`).MatchString(d.Get("package_file_source").(string)) {
-		if !strings.HasPrefix(localFilePath, os.TempDir()) {
-			log.Printf("[WARN] Refusing to remove file '%s' as it's not in the temporary directory: timestamp=%s",
-				localFilePath, time.Now().UTC().Format(time.RFC3339))
-			return diags
-		}
-
-		if err := os.Remove(localFilePath); err != nil {
-			log.Printf("[WARN] Failed to remove downloaded package file '%s': %v: timestamp=%s",
-				localFilePath, err, time.Now().UTC().Format(time.RFC3339))
-		} else {
-			log.Printf("[INFO] Successfully removed downloaded package file '%s': timestamp=%s",
-				localFilePath, time.Now().UTC().Format(time.RFC3339))
-		}
-	}
+	common.CleanupDownloadedPackage(d.Get("package_file_source").(string), localFilePath)
 
 	return append(diags, readNoCleanup(ctx, d, meta)...)
 }
